rename unaryerror to errorintercept
