+++ backend/middleware/errorintercept/errorintercept_test.go
@@ -1,4 +1,4 @@
-package unaryerror
+package errorintercept
 
 import (
 	"context"
